move to the proxy connection
